checkpolicy: Set errno to 0 before running strtol
Since C programs only set errno to 0 once at the start of the program, other function calls
above might have changed errno, whether intentionally or not. As this portion of the code
is designed to error out if errno is set to a non-zero value, it must be reset to 0 to
ensure that the strtol call is actually what's setting it to indicate an error.

See: https://www.securecoding.cert.org/confluence/pages/viewpage.action?pageId=6619179

Fixes cases where valid policyvers are treated as invalid, causing the program to exit.
(Example: Invalid policyvers specified: 26)

Change-Id: I3be2f66b1dc3d0a1b641682f7f486a076f594122